+++ util.js
@@ -403,6 +403,10 @@ navigateTo = function (url) {
             window.history.pushState(jsonurl, "", url);
             cache.put (jsonurl, data);
             updatePage (data);
+        }).fail(function (error) {
+            window.history.pushState(jsonurl, "", url);
+            $("#error").html (error.responseText);
+            console.log('getJSON failed: ' + JSON.stringify(error, null, 4));
         });
     }
 };
@@ -557,6 +561,9 @@ function initiateAjax (url, theme, templates) {
             window.history.replaceState(url, "", window.location);
         }
         handleLinks ();
+    }).fail(function (error) {
+        document.write (error.responseText);
+        console.log('getJSON failed: ' + JSON.stringify(error, null, 4));
     });
 }
 
@@ -585,6 +592,9 @@ function initiateTwig(url, theme, templates = 'templates') {
             window.history.replaceState(url, "", window.location);
         }
         handleLinks ();
+    }).fail(function (error) {
+        document.write (error.responseText);
+        console.log('getJSON failed: ' + JSON.stringify(error, null, 4));
     });
 }
  */